Move files over from PseudomodesTTEDOPA
Add the `closure.jl` file from the PseudomodesTTEDOPA package.
The previously called `Closure` object would have been named
`MarkovianClosure`, since this is what it represents (just `Closure`
would have been too generic), but this is also the package name and
Julia doesn't accept it, so it was called
`SemicircleMarkovianClosure`, given that it comes from a semicircle
spectral density. This object is not exposed to the final user
anyway, so this name is not very important.