Incorporate standard parameters in package
Hardcode the standard parameters of the semicircle Markovian closure,
i.e. the ones from a unit semicircle spectral density, directly in the
package itself, without requiring that the user provide them themselves.
This way MC objects can be constructed just by specifying the asymptotic
frequency and coupling constant and the number of pseudomodes.